requirements.txt
beautifulsoup4
cartopy
ipykernel
jupyter
matplotlib
notebook
numpy
pydap
python-dotenv
requests